Divide Aboutme.tsx
