+++ src/types/props.ts
@@ -1,3 +1,12 @@
+import { type } from "os"
+
 export type SiteDetailProps = {
   skills: string[];
+}
+
+export type AboutmeProps = {
+  items: {
+    title: string;
+    text: string;
+  }[];
 }
