+++ src/view/pages/Aboutme.tsx
@@ -1,6 +1,7 @@
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 import styles from './pages.module.css';
 import MediaQuery from "react-responsive";
+import AboutmeInfoes from '../templates/PC_Components/AboutmeInfoes';
 
 const Aboutme: FC<{}> = () => {
   
@@ -11,56 +12,15 @@ const Aboutme: FC<{}> = () => {
     {title: '年齢', text: '23歳：大学院、修士1年生'},
     {title: '言語・技術', text: 'React.js,  TypeScript,  Next.js,  Redux, ContextAPI'},
   ];
-
-  const [detail, setDetail] = useState({ title: items[0].title, text: items[0].text });
-
-  const detail_info = <div>
-                        <h1> {detail.title} </h1>
-                        <p> {detail.text} </p>
-                      </div>
-
-  const handleOnMouseEnter = (title: string, text: string): void => {
-    setDetail({ title: title, text: text });
-  }
   
   return(
     <div>
       <h1 className={styles.title}>About Me</h1>
         <MediaQuery minWidth={769}>
-          <ul className={styles.profiles}>
-            {items.map((temp, index) => (
-              index!==0 ?
-              <li key={index} className={styles.profiles_content} onMouseEnter={()=>handleOnMouseEnter(temp.title, temp.text)}>
-                <p>{temp.title}</p>
-              </li> : 
-              <li key={index} className={styles.profiles_content_last} onMouseEnter={()=>handleOnMouseEnter(temp.title, temp.text)}>
-                <p>{temp.title}</p>
-              </li>
-            ))
-            }
-          </ul>
-          
-          <hr></hr>
-          {detail_info}
-          <hr></hr>
+          <AboutmeInfoes items={items} />
         </MediaQuery>
         <MediaQuery query="(max-width: 768px)">
-          <ul className={styles.profiles_2}>
-            {items.map((temp, index) => (
-              index!==0 ?
-              <li key={index} className={styles.profiles_content_2} onMouseEnter={()=>handleOnMouseEnter(temp.title, temp.text)}>
-                <p>{temp.title}</p>
-              </li> : 
-              <li key={index} className={styles.profiles_content_last_2} onMouseEnter={()=>handleOnMouseEnter(temp.title, temp.text)}>
-                <p>{temp.title}</p>
-              </li>
-            ))
-            }
-          </ul>
-          
-          <hr></hr>
-          {detail_info}
-          <hr></hr>
+          <AboutmeInfoes items={items} />
         </MediaQuery>
 
     </div>
