+++ src/view/templates/PC_Components/pc.module.css
@@ -144,3 +144,83 @@
 }
 
 /*----------------- ContactInfoes.tsx ------------------ */
+
+/*----------------- Aboutme.tsx ------------------ */
+.profiles {
+  list-style: none;
+  margin: 0 auto;
+  padding: 0;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  width: 500px;
+}
+
+.profiles_2 {
+  list-style: none;
+  margin: 0 auto;
+  padding: 0;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  width: 250px;
+}
+
+.profiles_content {
+  font-weight: bold;
+  margin: 5px;
+  width: 200px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid black;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: 0.5s;
+}
+
+.profiles_content_2 {
+  font-weight: bold;
+  margin: 5px;
+  width: 100px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid black;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: 0.5s;
+}
+
+.profiles_content_last {
+  font-weight: bold;
+  margin: 5px;
+  width: 410px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid black;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: 0.5s;
+}
+
+.profiles_content_last_2 {
+  font-weight: bold;
+  margin: 5px;
+  width: 210px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid black;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: 0.5s;
+}
+
+.profiles li:hover {
+  transition: 0.5s;
+  background-color: #c782d8;
+}
+
+/*----------------- Aboutme.tsx ------------------ */
